# server/docs/swagger.yaml
definitions:
  dto.LoginDto:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  dto.TokenDto:
    properties:
      accessToken:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns access and refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: loginDto
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TokenDto'
      summary: User login
      tags:
      - auth
...
swagger: "2.0"